scripts/requirements.txt
progressbar2==3.39.3
python-chess==0.27.3
python-utils==2.3.0
six==1.12.0
